+++ changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Get error by input name.
+- Don't replace :field placeholder with input name in case no caption exists.
 
 ## [0.1.3] - 2022-04-14
 ### Changed